requirements.txt
flask
python-dotenv
eth-account
hyperliquid-python-sdk @ git+https://github.com/hyperliquid-dex/hyperliquid-python-sdk.git